added support for displaying splits on a run

+++ app/src/main/java/com/mcwilliams/theninjamethod/strava/model/activitydetail/SplitsMetric.kt
@@ -1,8 +1,8 @@
 package com.mcwilliams.theninjamethod.strava.model.activitydetail
 
 data class SplitsMetric(
-    val average_speed: Double,
-    val distance: Double,
+    val average_speed: Float,
+    val distance: Float,
     val elapsed_time: Int,
     val elevation_difference: Double,
     val moving_time: Int,
